+++ src/components/communityBoard/CommunityBoard-detail/Footer/type.ts
@@ -1,0 +1,4 @@
+export interface FooterProps {
+  postId: number;
+  fetchComments: () => void;
+}
